+++ ivy_tests/test_ivy/test_functional/test_experimental/test_core/test_manipulation.py
@@ -90,37 +90,6 @@ def test_moveaxis(
     )
 
 
-# ndenumerate
-@handle_test(
-    fn_tree="functional.ivy.experimental.ndenumerate",
-    dtype_and_x=helpers.dtype_and_values(
-        available_dtypes=helpers.get_dtypes("valid"),
-        min_num_dims=1,
-    ),
-)
-def test_ndenumerate(dtype_and_x):
-    values = dtype_and_x[1][0]
-    for (index1, x1), (index2, x2) in zip(
-        np.ndenumerate(values), ivy.ndenumerate(values)
-    ):
-        assert index1 == index2 and x1 == x2
-
-
-# ndindex
-@handle_test(
-    fn_tree="functional.ivy.experimental.ndindex",
-    dtype_x_shape=helpers.dtype_and_values(
-        available_dtypes=helpers.get_dtypes("valid"),
-        min_num_dims=1,
-        ret_shape=True,
-    ),
-)
-def test_ndindex(dtype_x_shape):
-    shape = dtype_x_shape[2]
-    for index1, index2 in zip(np.ndindex(shape), ivy.ndindex(shape)):
-        assert index1 == index2
-
-
 # heaviside
 @handle_test(
     fn_tree="functional.ivy.experimental.heaviside",
